ComboBox für die Serialisierung hinzugefügt, Save und Load Button hinzugefügt

+++ u1/src/main/java/helper/IDGenerator.java
@@ -1,14 +1,14 @@
 package helper;
 
 public class IDGenerator {
-	private static int IDmax = 999999;
-	int id = 0;
+	private static long IDmax = 999999;
+	long id = 0;
 	
 	public IDGenerator() {
 		
 	}
 	
-	public int getId() throws IDOverflow {
+	public long getId() throws IDOverflow {
 		if (this.id == IDmax) {
 			throw new IDOverflow();
 		}
